+++ packages/core-sdk/src/accounts/subgraph.ts
@@ -35,3 +35,64 @@ export async function getSellerByOperator(
 
   return sellers[0];
 }
+
+export const getSellerByAdminQuery = `
+query GetSellersByAdmin($admin: String!) {
+  sellers(where: {
+    admin: $admin
+  }) {
+    ...sellerFields
+  }
+}
+${sellerFieldsFragment}
+`;
+
+export async function getSellerByAdmin(
+  subgraphUrl: string,
+  adminAddress: string
+): Promise<RawSellerFromSubgraph> {
+  const { sellers } = await fetchSubgraph<{ sellers: RawSellerFromSubgraph }>(
+    subgraphUrl,
+    getSellerByOperatorQuery,
+    { admin: adminAddress }
+  );
+
+  return sellers[0];
+}
+
+export const getSellerByClerkQuery = `
+query GetSellersByOperator($clerk: String!) {
+  sellers(where: {
+    clerk: $clerk
+  }) {
+    ...sellerFields
+  }
+}
+${sellerFieldsFragment}
+`;
+
+export async function getSellerByClerk(
+  subgraphUrl: string,
+  clerkAddress: string
+): Promise<RawSellerFromSubgraph> {
+  const { sellers } = await fetchSubgraph<{ sellers: RawSellerFromSubgraph }>(
+    subgraphUrl,
+    getSellerByOperatorQuery,
+    { clerk: clerkAddress }
+  );
+
+  return sellers[0];
+}
+
+export async function getSellerByAddress(
+  subgraphUrl: string,
+  address: string
+): Promise<RawSellerFromSubgraph> {
+  const [operator, admin, clerk] = await Promise.all([
+    getSellerByOperator(subgraphUrl, address),
+    getSellerByAdmin(subgraphUrl, address),
+    getSellerByClerk(subgraphUrl, address)
+  ]);
+
+  return operator || admin || clerk;
+}