+++ packages/example-parent-app/src/views/manage/index.tsx
@@ -32,7 +32,17 @@ export function Manage() {
     if (offer) {
       const el = document.createElement("div");
       widgetRef.current.appendChild(el);
-      manageOffer(offer.id, CONFIG, el);
+      manageOffer(
+        offer.id,
+        {
+          ...CONFIG,
+          widgetsUrl: "http://localhost:3000"
+        },
+        el,
+        {
+          forceBuyerView: false
+        }
+      );
 
       return () => el.remove();
     }