+++ packages/ipfs-storage/src/ipfs.ts
@@ -1,15 +1,11 @@
-import { utils, MetadataStorage, AnyMetadata } from "@bosonprotocol/common";
+import { MetadataStorage, AnyMetadata } from "@bosonprotocol/metadata";
 import { create, IPFSHTTPClient, Options } from "ipfs-http-client";
 import fetch from "cross-fetch";
 import { concat, toString } from "uint8arrays";
 import { CID } from "multiformats/cid";
 import { validateMetadata } from "./validation";
 import { DEFAULT_THE_GRAPH_IPFS_URL } from "./constants";
-import {
-  convertFromERC721Metadata,
-  convertToERC721Metadata,
-  ERC721Metadata
-} from "./utils";
+import { convertToERC721Metadata, ERC721Metadata, sortObjKeys } from "./utils";
 
 export class IpfsMetadata implements MetadataStorage {
   public ipfsClient: IPFSHTTPClient;
@@ -27,18 +23,13 @@ export class IpfsMetadata implements MetadataStorage {
   public async storeMetadata(metadata: AnyMetadata): Promise<string> {
     validateMetadata(metadata);
     const metadataConformingToERC721 = convertToERC721Metadata(metadata);
-    const metadataWithSortedKeys = utils.metadata.sortObjKeys(
-      metadataConformingToERC721
-    );
+    const metadataWithSortedKeys = sortObjKeys(metadataConformingToERC721);
     const cid = await this.add(metadataWithSortedKeys);
     return cid;
   }
 
   public async getMetadata(metadataUriOrHash: string): Promise<AnyMetadata> {
-    const metadataConformingToERC721 = await this.get<ERC721Metadata>(
-      metadataUriOrHash
-    );
-    const metadata = convertFromERC721Metadata(metadataConformingToERC721);
+    const metadata = await this.get<ERC721Metadata>(metadataUriOrHash);
     validateMetadata(metadata);
 
     return metadata;