+++ packages/metadata/jest.config.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const baseConfig = require("../../jest.config.sdk");
+
+module.exports = baseConfig;