+++ packages/example-parent-app/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/example-parent-app@1.6.0...@bosonprotocol/example-parent-app@1.7.0) (2022-05-27)
+
+### Features
+
+* migrate to new testing env ([#93](https://github.com/bosonprotocol/core-components/issues/93)) ([5af0e2f](https://github.com/bosonprotocol/core-components/commit/5af0e2f5c4ab3c90a80b0ce323464c53690c7dc1))
+
+# [1.6.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/example-parent-app@1.5.0...@bosonprotocol/example-parent-app@1.6.0) (2022-05-13)
+
+### Features
+
+* wallet change notificaiton and forcable buyer view ([#89](https://github.com/bosonprotocol/core-components/issues/89)) ([ff321e7](https://github.com/bosonprotocol/core-components/commit/ff321e7cc82406906dac8fa531c41a1263c9696c))
+* **widgets:** enforce correct chain in wallet ([#88](https://github.com/bosonprotocol/core-components/issues/88)) ([d9060b3](https://github.com/bosonprotocol/core-components/commit/d9060b3c361ab168b0f7b4c319e03f322ebc6bab))
+
+# [1.5.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/example-parent-app@1.4.0...@bosonprotocol/example-parent-app@1.5.0) (2022-05-10)
+
+### Features
+
+* separate metadata package ([#86](https://github.com/bosonprotocol/core-components/issues/86)) ([2bf5069](https://github.com/bosonprotocol/core-components/commit/2bf5069256592e8ed5e80a3e557e1402ba437fc9))
+
+# [1.4.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/example-parent-app@1.3.0...@bosonprotocol/example-parent-app@1.4.0) (2022-05-09)
+
+### Features
+
+* add commit widget ([#85](https://github.com/bosonprotocol/core-components/issues/85)) ([4aec8ce](https://github.com/bosonprotocol/core-components/commit/4aec8cec60d07588f356da0997ec4414f22a5b55))
+
 # [1.3.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/example-parent-app@1.2.0...@bosonprotocol/example-parent-app@1.3.0) (2022-05-04)
 
 ### Features